Refactor Link to multiple files and add types [WIP]
