+++ packages/components/link/index.tsx
@@ -1,129 +1,8 @@
 import React, { MouseEvent, useEffect, useRef, useCallback } from "react";
 import { warn, connect, useConnect } from "frontity";
 import useInView from "@frontity/hooks/use-in-view";
-import { Package } from "frontity/types";
-
-const toPrefetch = new Set<string>();
-let isProcessingQueue = false;
-
-export const config = {
-  hoverDelay: 50,
-  requestsPerBatch: 4,
-  batchInterval: 1000,
-};
-
-/**
- * Process the queue of links to prefetch.
- *
- * @param prefetcher The function used to prefetch the link.
- */
-function processQueue(prefetcher: (link: string) => void) {
-  let batchInterval;
-
-  /**
-   * Process a batch of prefetches
-   */
-  const process = () => {
-    const batch = Array.from(toPrefetch).slice(0, config.requestsPerBatch);
-
-    // if batch is empty, stop process and allow it to run again if necessary.
-    if (batch.length === 0) {
-      isProcessingQueue = false;
-      clearTimeout(batchInterval);
-      return;
-    }
-
-    batch.forEach((link) => {
-      prefetcher(link);
-      toPrefetch.delete(link);
-    });
-  };
-
-  batchInterval = setInterval(() => {
-    process();
-  }, config.batchInterval);
-}
-
-/**
- * Executes the callback when the hover event is triggered for the element.
- *
- * @param el The element to watch for hover events.
- * @param cb The callback that should run should an hover event happen.
- */
-function onHover(el: HTMLAnchorElement, cb: () => void) {
-  let timeout;
-
-  const mouseOverHandler = () => {
-    timeout = setTimeout(() => {
-      cb();
-    }, config.hoverDelay);
-  };
-
-  const mouseOutHandler = () => {
-    clearTimeout(timeout);
-  };
-
-  el.addEventListener("mouseover", mouseOverHandler);
-  el.addEventListener("mouseout", mouseOutHandler);
-
-  return () => {
-    el.removeEventListener("mouseover", mouseOverHandler);
-    el.removeEventListener("mouseout", mouseOutHandler);
-  };
-}
-
-/**
- * Props for React component {@link Link}.
- */
-interface LinkProps {
-  /**
-   * The URL to link to.
-   */
-  link: string;
-
-  /**
-   * The target of the anchor:
-   * https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a#target.
-   *
-   * @defaultValue "_self"
-   */
-  target?: "_self" | "_blank";
-
-  /**
-   * The onClick handler. Can be used to pass an optional callback that will be
-   * invoked on click.
-   */
-  onClick?: (event: MouseEvent<HTMLAnchorElement>) => void;
-
-  /**
-   * Whether the browser should scroll up to the top upon navigating to a new
-   * page.
-   *
-   * @defaultValue true
-   */
-  scroll?: boolean;
-
-  /**
-   * Whether frontity should automatically prefetch this link or not.
-   * The prefetching mode is controlled through state.theme.prefetch.
-   *
-   * @defaultValue true
-   */
-  prefetch?: boolean;
-
-  /**
-   * Indicates the element that represents the current item within a container
-   * or set of related elements:
-   * https://www.w3.org/TR/wai-aria-1.1/#aria-current.
-   */
-  "aria-current"?: React.AriaAttributes["aria-current"];
-
-  /**
-   * Represents any other prop that can be passed to Link.
-   * These props are passed down to the `<a/>` element.
-   */
-  [key: string]: any;
-}
+import { isProcessingQueue, toPrefetch, onHover, processQueue } from "./utils";
+import { Packages, LinkProps, NavigatorWithConnection } from "./types";
 
 /**
  * The Link component that enables linking to internal pages in a frontity app.
@@ -152,10 +31,9 @@ const Link: React.FC<LinkProps> = ({
   target = "_self",
   scroll = true,
   prefetch = true,
-  "aria-current": ariaCurrent,
   ...anchorProps
 }) => {
-  const { state, actions } = useConnect<Package>();
+  const { state, actions } = useConnect<Packages>();
   const { ref: inViewRef, inView } = useInView({
     triggerOnce: true,
     rootMargin: "200px",
@@ -181,8 +59,11 @@ const Link: React.FC<LinkProps> = ({
   const isExternal = link.startsWith("http");
 
   useEffect(() => {
-    // Checks if user is on slow connection or has enabled data saver
-    const _navigator = navigator as Navigator & { connection };
+    /**
+     * Checks if user is on slow connection or has enabled data saver.
+     */
+    const _navigator = window.navigator as NavigatorWithConnection;
+
     const isSlowConnection =
       _navigator?.connection?.saveData ||
       (_navigator?.connection?.effectiveType || "").includes("2g");
@@ -192,12 +73,12 @@ const Link: React.FC<LinkProps> = ({
     }
 
     /**
-     * Prefetches the link only if necessary
+     * Prefetches the link only if necessary.
      *
-     * @param link The link to prefetch
-     * @param runNow Whether the prefetch should be executed immediatelly.
+     * @param link - The link to prefetch.
+     * @param runNow - Whether the prefetch should be executed immediatelly.
      */
-    const maybePrefetch = (link: string, runNow: boolean = false) => {
+    const maybePrefetch = (link: string, runNow = false) => {
       if (toPrefetch.has(link)) {
         return;
       }
